:sparkles: New commands for priwo's cli.
This introduces some new commands into priwo's cli, including `max` for getting the maximum size and `available` for getting a table of all the data formats available in priwo. These are also available as functions for the user to access, so that they may query this info directly in the REPl as well.